+++ src/_genFunc.c
@@ -32,7 +32,7 @@ void swapi(int *val1, int *val2){
 //return 0 if sorted correctly
 //return 1 if sorted incorrectly
 //mode if mode is set to 1 it also prints out a info message to the console, if it is 0 only return value
-int checkSort(int *arr, int size, short int mode){
+int checkSort(int *arr, int size, int mode){
 
     int tempLeftToComapare = 0;
     int largerElementFound = 0;
